Switch to KotlinCompilerPluginSupportPlugin
