--- src/gradle-plugin/src/main/kotlin/com/josephdwyer/katana/gradle/Katana.kt
@@ -1,12 +1,0 @@
-package com.josephdwyer.katana.gradle
-
-import org.gradle.api.Plugin
-import org.gradle.api.Project
-
-class Katana : Plugin<Project> {
-
-    override fun apply(project: Project) {
-        project.extensions.create("katana", KatanaPluginExtension::class.java)
-    }
-
-}
