+++ src/gradle-plugin/src/main/kotlin/com/josephdwyer/katana/gradle/KatanaPluginExtension.kt
@@ -1,6 +1,8 @@
 package com.josephdwyer.katana.gradle
 
-open class KatanaPluginExtension {
+import org.gradle.api.Project
+
+open class KatanaPluginExtension(project: Project) {
     /**
         The (fully qualified) file path to write the Katana data to, should be a JSON file
      */
